wordpan/ai/pyproject.toml
[project]
name = "wordpan-ai"
version = "0.0.1"
description = "Wordpan AI services"
classifiers = ["Private :: Do Not Upload"]
authors = []
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.9.3",
    "arize-phoenix==12.4.0",
    "crewai==0.201.0",
    "crewai-tools==0.75.0",
    "flask[async]>=3.1.0",
    "flask-cors>=6.0.1",
    "openinference-instrumentation-crewai==0.1.10",
    "openinference-instrumentation-litellm==0.1.23",
    "opentelemetry-exporter-otlp>=1.34.1",
    "opentelemetry-sdk>=1.34.1",
    "pydantic>=1.8.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.22.0",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.13",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-mock>=3.14.1",
    "ruff>=0.12.1",
    "watchdog>=6.0.0",
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest]
addopts = "-ras -l -vv"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 120
target-version = "py313"
